+++ Philas/app/Model/Entity/User.php
@@ -30,7 +30,10 @@ class User {
   public $cpf;
 
   /** @var string Tipo relacionado as permissoões do usuário */
-  public $tipo = 'Comum';
+  public $tipo;
+
+  /** @var array Tipos possíveis para os usuários*/
+  public static $tipos = ['admin' => 'Admin', 'default' => 'Comum'];
 
   /** @var string Tabela atual no banco de dados da Entidade */
   private static $table = "usuario";
@@ -86,7 +89,7 @@ class User {
       'email'    => $this->email,
       'telefone' => $this->telefone,
       'cpf'      => $this->cpf,
-      'tipo'     => $this->tipo
+      'tipo'     => $this->tipo ?? self::$tipos['default']
     ]);
 
     // SUCESSO
@@ -99,16 +102,21 @@ class User {
    * @return  boolean
    */
   public function update() {
-    // ATUALIZA O USUÁRIO NO BANCO DE DADOS
-    return (new Database(self::$table))->update('id = ' . $this->id, [
+    // DADOS DO USUÁRIO
+    $userData = [
       'nome'     => $this->nome,
       'login'    => $this->login,
-      'senha'    => password_hash($this->senha, PASSWORD_DEFAULT),
       'email'    => $this->email,
       'telefone' => $this->telefone,
       'cpf'      => $this->cpf,
-      'tipo'     => $this->tipo
-    ]);
+      'tipo'     => $this->tipo ?? self::$tipos['default']
+    ];
+
+    // SE NÃO HOUVER SENHA, IGNORE-A
+    if (isset($this->senha)) $userData['senha'] = password_hash($this->senha, PASSWORD_DEFAULT);
+
+    // ATUALIZA O USUÁRIO NO BANCO DE DADOS
+    return (new Database(self::$table))->update('id = ' . $this->id, $userData);
   }
 
   /**
@@ -117,8 +125,10 @@ class User {
    * @return  boolean
    */
   public function delete() {
-    // EXCLUI O USUÁRIO DO BANCO DE DADOS
-    return (new Database(self::$table))->delete('id = ' . $this->id);
+    // ALTERA OS CAMPOS COM CHAVE ESTRANGEIRA
+    if (Schedule::deleteUserForeignKeys($this))
+      // EXCLUI O USUÁRIO DO BANCO DE DADOS
+      return (new Database(self::$table))->delete('id = ' . $this->id);
   }
 
   /**
